change meta-data storage interface
this was always racy for fs, because the filelock was
applied separately for load and store. So every rmw-cycle
when updating could corrupt any other change.

This was especially stupid given the API where the
in-memory update of our meta-data happened in the
function retrieving apples' meta-data :-D i.e. if we
want to make it safe, we'd have to keep the lock open
during the download of the meta-data.

But yeah, known issue, defer to later.

But now, with a second storage option (GCS), it would
make things more complicated. So we switch to a pure
retrival function, do the rmw-cycle in the save/store
function. This keeps the filelock short in the fs store
and allows for an optimistic lock with GCS, because we
just retry with the retrieved meta-data.