Add timeout to OTA mirror
This allows us to do an ordered shutdown of the mirroring processes before the actual GHA workflow timeout hits us.

If that happens our workflow will be notified as "Failed" although running into that timeout is totally accepted and not an actual error state.

If we have an application-level timeout we can do an orderly shutdown at the most opportune point and never hit the workflow limit.