Replace uncaught meta-download RuntimeErrors with sentry error logs
Everytime the meta download fails it is due to an intermittent issue like a 403. There is no reason for the mirror workflow to stop, this should just be logged, so we know that it didn't run successfully.

Apparently it is able to download for other devices, so maybe later on we just retry a couple of times.